# docs/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: Cybertech E-Commerce API
  description: |
    Backend REST API for the Cybertech e-commerce platform.

    Covers the full e-commerce lifecycle: product catalog, search,
    cart, wishlist, orders, payments (Stripe), reviews and moderation,
    bank cards (PCI-aware), discount campaigns, user management
    (Keycloak-backed) and behavioural event ingestion.

    Authentication is performed via Keycloak-issued JWT bearer tokens.
    Public endpoints (catalog browsing, registration, Stripe webhook)
    are explicitly whitelisted; every other endpoint requires a
    valid Bearer token and may further restrict to USER or ADMIN roles.
  contact:
    name: Cybertech Team
    email: dev@cybertech.local
  license:
    name: Proprietary
    url: https://cybertech.local/license
  version: "1.0"
servers:
- url: http://localhost:8081
  description: Local dev
- url: https://api.cybertech.local
  description: Production
security:
- keycloak: []
...
paths:
...
  /api/v1/services/management/order/update:
    post:
      tags:
      - Order
      summary: Update an existing Order by UUID
      description: |
        Updates an existing order's details based on their unique UUID. Fields not provided will not be updated.
      operationId: updateOrder
      parameters:
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        description: Order data for update. Only provide fields that need to be changed.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderUpdateRequestDto"
        required: true
      responses:
        "200":
          description: Order updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponseDto"
        "400":
          description: Invalid input data / Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: Internal server error during order update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
  /api/v1/services/management/order/retry-payment/{uuid}:
    post:
      tags:
      - Order
      summary: Retry payment for a failed order
      description: Retries the payment process for an order that is in PAYMENT_FAILED
        status. Checks stock availability before proceeding.
      operationId: retryPayment
      parameters:
      - name: uuid
        in: path
        description: The UUID of the order to retry payment for
        required: true
        schema:
          type: string
          format: uuid
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
          - 1.0.0
      responses:
        "200":
          description: Payment retried successfully (Order Paid)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponseDto"
        "400":
          description: Invalid request or Order not in FAILED state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
  /api/v1/services/management/order/place:
    post:
      tags:
      - Order
      summary: Place a new order (authenticated users only)
      description: |
        Creates a new order for the authenticated user.
        The request must include the list of products to purchase, quantities,
        shipping address, and payment type.
        Requires a valid JWT token in the Authorization header (Bearer token).
      operationId: placeOrder
      parameters:
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        description: Order creation data for the currently authenticated user.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderPlacingRequestDto"
        required: true
      responses:
        "201":
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponseDto"
        "400":
          description: Invalid request or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden - user not allowed to perform this operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: Internal server error while placing the order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
  /api/v1/services/management/order/cancel:
    post:
      tags:
      - Order
      summary: Cancel an existing order (authenticated users only)
      description: |
        Cancels an order identified by its UUID.
        The operation is allowed only if the order belongs to the authenticated user and is in a cancellable state.
        Requires a valid JWT token in the Authorization header (Bearer token).
      operationId: cancelOrder
      parameters:
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderCancellationRequestDto"
        required: true
      responses:
        "200":
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponseDto"
        "400":
          description: Invalid order UUID or order cannot be cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Forbidden - user not allowed to cancel this order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: Internal server error while cancelling the order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
...
  /api/v1/services/management/order/status/{uuid}:
    get:
      tags:
      - Order
      summary: Get the current status of an order by UUID
      description: Lightweight status read for the order-confirmation polling loop
        after a Stripe payment. Ownership-checked at the service layer.
      operationId: getOrderStatusByUuid
      parameters:
      - name: uuid
        in: path
        description: UUID of the order to read status for
        required: true
        schema:
          type: string
          format: uuid
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
          - 1.0.0
      responses:
        "200":
          description: Order status retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderStatusDto"
        "401":
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Operation forbidden - order does not belong to the authenticated
            user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
  /api/v1/services/management/order/mine:
    get:
      tags:
      - Order
      summary: List the authenticated user's orders (paginated)
      description: |
        Frontend-gap #1 — paginated listing of every order belonging to the authenticated
        user, optionally filtered by status. The user's identity is resolved from the JWT
        subject so there is no IDOR risk: callers see only their own orders.
      operationId: getMyOrders
      parameters:
      - name: status
        in: query
        description: Optional set of OrderStatus values to keep — repeat the query
          param to pass multiple values
        required: false
        schema:
          type: array
          items:
            type: string
            enum:
            - CREATED
            - AWAITING_PAYMENT
            - PAYMENT_FAILED
            - PAID
            - AWAITING_SHIPPING
            - SHIPPED
            - DELIVERED
            - RETURNED
            - CANCELED
            - REFUNDED
          uniqueItems: true
      - name: page
        in: query
        description: 0-based page index (default 0)
      - name: size
        in: query
        description: page size (default 20)
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
          - 1.0.0
      responses:
        "200":
          description: Page of orders for the caller
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Page"
        "401":
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
      security:
      - keycloak: []
  /api/v1/services/management/order/get/{uuid}:
    get:
      tags:
      - Order
      summary: Request a Order by UUID
      description: Fetches a Order's details based on their unique UUID.
      operationId: getOrderByUuid
      parameters:
      - name: uuid
        in: path
        description: UUID for searching a Order
        required: true
        schema:
          type: string
          format: uuid
      - name: X-API-VERSION
        in: header
        schema:
          type: string
          default: 1.0.0
          enum:
          - "1.0"
          - 1.0.0
      responses:
        "200":
          description: Order found successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderResponseDto"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "401":
          description: Unauthorized - JWT token missing or invalid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "403":
          description: Operation forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponseDto"
...
components:
  schemas:
...
    ErrorResponseDto:
      type: object
      properties:
        message:
          type: string
        httpStatusCode:
          type: integer
          format: int32
        errorCodeType:
          type: string
          enum:
          - TECHNICAL
          - FUNCTIONAL
...
    OrderItemCreateRequestDto:
      type: object
      properties:
        productUuid:
          type: string
          format: uuid
        quantity:
          type: integer
          format: int32
          minimum: 1
      required:
      - productUuid
      - quantity
    OrderUpdateRequestDto:
      type: object
      description: DTO for updating an existing Order. Only provide fields that need
        to be changed.
      properties:
        uuid:
          type: string
          format: uuid
        paymentType:
          type: string
          enum:
          - MASTERCARD
          - VISA
          - APPLE_PAY
          - GOOGLE_PAY
        shippingType:
          type: string
          enum:
          - STANDARD
          - EXPRESS
        shippingProvider:
          type: string
          enum:
          - DHL
          - FEDEX
        shippingStreet:
          type: string
          minLength: 1
        shippingCity:
          type: string
          minLength: 1
        shippingZipCode:
          type: string
          minLength: 1
        shippingCountry:
          type: string
          minLength: 1
        itemUpdateRequestDtoList:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemCreateRequestDto"
          minItems: 1
      required:
      - itemUpdateRequestDtoList
      - paymentType
      - shippingCity
      - shippingCountry
      - shippingProvider
      - shippingStreet
      - shippingType
      - shippingZipCode
      - uuid
    OrderItemResponseDto:
      type: object
      properties:
        orderItemUuid:
          type: string
          format: uuid
        productUuid:
          type: string
          format: uuid
        productName:
          type: string
        quantity:
          type: integer
          format: int32
        unitPrice:
          type: number
        lineItemTotalPrice:
          type: number
    OrderResponseDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        userUuid:
          type: string
          format: uuid
        orderDate:
          type: string
          format: date
        status:
          type: string
          enum:
          - CREATED
          - AWAITING_PAYMENT
          - PAYMENT_FAILED
          - PAID
          - AWAITING_SHIPPING
          - SHIPPED
          - DELIVERED
          - RETURNED
          - CANCELED
          - REFUNDED
        totalAmount:
          type: number
        refundedAmount:
          type: number
        shippingAddress:
          type: string
        orderItems:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemResponseDto"
    OrderPlacingRequestDto:
      type: object
      description: OrderPlacingRequestDto for placing order
      properties:
        paymentType:
          type: string
          enum:
          - MASTERCARD
          - VISA
          - APPLE_PAY
          - GOOGLE_PAY
        shippingType:
          type: string
          enum:
          - STANDARD
          - EXPRESS
        shippingProvider:
          type: string
          enum:
          - DHL
          - FEDEX
        shippingStreet:
          type: string
          minLength: 1
        shippingCity:
          type: string
          minLength: 1
        shippingZipCode:
          type: string
          minLength: 1
        shippingCountry:
          type: string
          minLength: 1
        discountKey:
          type: string
          minLength: 1
      required:
      - discountKey
      - paymentType
      - shippingCity
      - shippingCountry
      - shippingProvider
      - shippingStreet
      - shippingType
      - shippingZipCode
    OrderCancellationRequestDto:
      type: object
      description: DTO containing the UUID of the order to cancel
      properties:
        orderUuid:
          type: string
          format: uuid
      required:
      - orderUuid
...
    OrderStatusDto:
      type: object
      properties:
        uuid:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - CREATED
          - AWAITING_PAYMENT
          - PAYMENT_FAILED
          - PAID
          - AWAITING_SHIPPING
          - SHIPPED
          - DELIVERED
          - RETURNED
          - CANCELED
          - REFUNDED
    Page:
      type: object
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        content:
          type: array
          items: {}
        number:
          type: integer
          format: int32
        first:
          type: boolean
        last:
          type: boolean
        numberOfElements:
          type: integer
          format: int32
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        empty:
          type: boolean
    PageableObject:
      type: object
      properties:
        offset:
          type: integer
          format: int64
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: boolean
        pageSize:
          type: integer
          format: int32
        pageNumber:
          type: integer
          format: int32
        unpaged:
          type: boolean
    SortObject:
      type: object
      properties:
        empty:
          type: boolean
        unsorted:
          type: boolean
        sorted:
          type: boolean
...
  securitySchemes:
    keycloak:
      type: http
      description: "Keycloak-issued JWT access token. Obtain one from the realm token\
        \ endpoint and pass it as the Authorization header: 'Bearer <token>'."
      scheme: bearer
      bearerFormat: JWT